Removed Helper class and moved its code into Utils
